+++ NEWS.md
@@ -1,3 +1,13 @@
+
+# BiocProject 0.2.1 - 2019-10-15
+
+## Added
+* added vignette: "Using BiocProject with tximeta"
+
+## Changed 
+* `subproject` argument to `amendments` in `BiocProject` function 
+* `Project` is not added as a first element in `Annotated@metadata`, it is appended to any existing ones
+
 # BiocProject 0.2 - 2019-04-19
 
 ## Added